Expose line and column number of parsed elements
The `linenum` and `colnum` flags were introduced to mark fields that
will save the line number and column number of an element.

The field should not really map to any field names on the YAML document,
since the parser won't really save it. Once line or column number is
saved to a field that carries the flags mentioned here, the parser is
done with that field.

Might fix https://github.com/go-yaml/yaml/issues/108