Tighten restrictions on float checks
ParseFloat() accepts strings that contain digits with a single 'e'
character somewhere in the middle as valid floats. The YAML spec does
not accept these.

This causes problems especially when dealing with short commit hashes,
e.g. `123456e1`

This regex isn't exactly the same as the YAML spec. It continues to
allow decimals like `0.1` to be parsed as floats.