Preserve JSON Codec Compatibility
This patch adds the function `PreserveJSONCodecCompatibility` in order
to provide the ability to preserve the ability take data unmarshaled
using this package and marshal it back to JSON using the Golang StdLib
JSON library.

The Go StdLib JSON marshaller cannot encode maps with keys other
than of type string. Therefore, maps created by this package's decoder
must use a string as the key-type in order for the same maps to be
marshaled into JSON at a later date.