omitempty use IsZero
As per issue #244, time.Time does not work well with
omitempty, because it has no public fields.

We make the omitempty logic consult the IsZero
method, if available, to determine whether to omit
a value.

Omitting objects with private fields seems a bit
dubious, but since that's existing behaviour,
we'll keep it.

We also make *time.Time work the same as time.Time
for the purposes of timestamp tags.

Fixes issue #244.