Allow specifying flow or block style at encode time
It is often useful to be able to encode a given value in both block and
flow style. The current struct tag-based API requires committing to
either block or flow style for a given field. It also provides no means
of requesting that the top-level struct use flow style.

Introduce a new API, Encoder.UseStyle, which requests that all encoded
values use a particular encoding style.