Fix -0b on 32-bit systems
7b8fd2d introduced a regression on 32-bit systems where -0b values would be
cast to an Int before being returned as a Int64, meaning that on 32-bit
systems, any negative value smaller than MinInt32 would get wrapped around and
returned incorrectly, causing the int64_neg_base2 decode_test to fail.

The test case also causes the bin test to fail on 32-bit systems because it
presumes the 64-bit edge case value can fit within an Int, which is not the
case on a 32-bit system.

Move the newly-introduced test case to the Int64-specific test cases, and
remove the undesirable coersion of Int64 values to Int when they do not fit.