Preserve data in existing fields that are not decoded.
This behavior change means it's now possible to pre-initialize
values and then provide them to decoding, with the result being
a sum of the previous data and the successfuly decoded fields.

Particular care needs to be taken in loops that decode into the same
value repeatedly. If the variable decoded into is declared in a scope
out of the loop block, following iterations will observe data decoded in
previous iterations. To fix that just move the variable declaration
inside the loop.

Discussed in #395.