Extend Node with start index for all types and end markers for scalars.
Copies additional marker fields from the parser's events to the relevant Node
instance to expose additional source locations in the public API.

Index is populated form the start event for all node types. Unlike Line/Column,
Index is a zero-based (rune) offset from the start of the YAML stream.

LineEnd, ColumnEnd and IndexEnd are populated from the end event. The fields
have the same semantics as the equivalent start marker fields.

For now, *End locations are only set/meaningful for scalars because end marker
info at the parser level seems spurious in some situations. However, the main
use case for exposing end markers is to make in-place edits to YAML source, and
scalars are the most likely targets, so this seems like a reasonable compromise
and therefore a sensible first step.