Fix indents of the foot comments going after the map
Removed the function that was adding the comments to the last key from
the map and replaced that with the separate comments queue.

Unfold that comment queue each time we get the end of a flow, a document and
a block.
Append comments to the end token if their indents are >= than the first
comment indent (this is safe as we only add comments to that queue if we
detect that their indent is < than the last key indent).

Additionally fixed the bug in the emitter, related to skipping
foot_comments for the mapping end events.

Fixes: https://github.com/go-yaml/yaml/issues/709

Signed-off-by: Artem Chernyshev <artem.0xD2@gmail.com>