Allow for custom struct tags such as `json`.
This is useful when using structs from 3rd-party packages which almost certainly only have `json` tags and not `yaml` tags.

This implementation does not change the default behavior for only looking for `yaml` tags, but allows the user to specify a list of tags to consider such as []string{"yaml", "json"} in an encoder or decoder.